api/requirements.txt
fastapi==0.110.1
mangum==0.17.0
motor==3.3.2
pydantic==2.6.4
python-dotenv==1.0.1
